[Observability Onboarding] Make custom cards always appear in search results (#208913)
Closes #207734

This PR:
* Changes the behavior of the search form to always include quickstart
flows in search results even when a category is not selected
* Refactors and cleans up the code a bit

https://github.com/user-attachments/assets/e5de7092-2d9f-41be-8d69-25954e5e4bff

## How to test

Make sure that the search works as expected and when clicking on the
cards it leads to the right places.

(cherry picked from commit 015a4ac618d08a1cf799c9004faaa02a62ba1b74)
