+++ x-pack/solutions/observability/plugins/observability_onboarding/public/application/onboarding_flow_form/onboarding_flow_form.tsx
@@ -26,12 +26,13 @@ import { css } from '@emotion/react';
 
 import { useSearchParams } from 'react-router-dom-v5-compat';
 import { useKibana } from '@kbn/kibana-react-plugin/public';
-import { OnboardingFlowPackageList } from '../packages_list';
+import { IntegrationCardItem } from '@kbn/fleet-plugin/public';
+import { PackageListSearchForm } from '../package_list_search_form/package_list_search_form';
 import { Category } from './types';
-import { useCustomCardsForCategory } from './use_custom_cards_for_category';
-import { useVirtualSearchResults } from './use_virtual_search_results';
+import { useCustomCards } from './use_custom_cards';
 import { LogoIcon, SupportedLogo } from '../shared/logo_icon';
 import { ObservabilityOnboardingAppServices } from '../..';
+import { PackageList } from '../package_list/package_list';
 
 interface UseCaseOption {
   id: Category;
@@ -147,11 +148,25 @@ export const OnboardingFlowForm: FunctionComponent = () => {
     [] // eslint-disable-line react-hooks/exhaustive-deps
   );
 
-  const customCards = useCustomCardsForCategory(
-    createCollectionCardHandler,
-    searchParams.get('category') as Category | null
-  );
-  const virtualSearchResults = useVirtualSearchResults();
+  const featuredCardsForCategoryMap: Record<Category, string[]> = {
+    host: ['auto-detect-logs', 'otel-logs'],
+    kubernetes: ['kubernetes-quick-start', 'otel-kubernetes'],
+    application: ['apm-virtual', 'otel-virtual', 'synthetics-virtual'],
+    cloud: ['azure-logs-virtual', 'aws-logs-virtual', 'gcp-logs-virtual'],
+  };
+  const customCards = useCustomCards(createCollectionCardHandler);
+  const featuredCardsForCategory: IntegrationCardItem[] = customCards.filter((card) => {
+    const category = searchParams.get('category') as Category;
+
+    if (category === null) {
+      return false;
+    }
+
+    const cardList = featuredCardsForCategoryMap[category] ?? [];
+
+    return cardList.includes(card.id);
+  });
+
   /**
    * Cloud deployments have the new Firehose quick start
    * flow enabled, so the ond card 'epr:awsfirehose' should
@@ -272,7 +287,7 @@ export const OnboardingFlowForm: FunctionComponent = () => {
       </EuiFlexGrid>
       {/* Hiding element instead of not rending these elements in order to preload available packages on page load */}
       <div
-        hidden={!searchParams.get('category') || !customCards}
+        hidden={featuredCardsForCategory.length === 0}
         role="group"
         aria-labelledby={packageListTitleId}
       >
@@ -310,11 +325,7 @@ export const OnboardingFlowForm: FunctionComponent = () => {
             </strong>
           </EuiTitle>
           <EuiSpacer size="m" />
-          <OnboardingFlowPackageList
-            customCards={customCards}
-            flowSearch={integrationSearch}
-            flowCategory={searchParams.get('category')}
-          />
+          <PackageList list={featuredCardsForCategory} />
         </div>
       </div>
 
@@ -329,20 +340,12 @@ export const OnboardingFlowForm: FunctionComponent = () => {
           </strong>
         </EuiText>
         <EuiSpacer size="m" />
-        <OnboardingFlowPackageList
-          showSearchBar={true}
+        <PackageListSearchForm
           searchQuery={integrationSearch}
-          flowSearch={integrationSearch}
           setSearchQuery={setIntegrationSearch}
           flowCategory={searchParams.get('category')}
-          customCards={(customCards || [])
-            .filter(
-              // Filter out collection cards and regular integrations that show up via search anyway
-              (card) => card.type === 'virtual' && !card.isCollectionCard
-            )
-            .concat(virtualSearchResults)}
+          customCards={customCards.filter((card) => !card.isCollectionCard)}
           excludePackageIdList={searchExcludePackageIdList}
-          joinCardLists
         />
       </div>
     </EuiPanel>
