+++ x-pack/solutions/observability/plugins/observability_onboarding/public/application/package_list_search_form/use_card_url_rewrite.test.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { addPathParamToUrl, toOnboardingPath } from './use_integration_card_list';
+import { addPathParamToUrl, toOnboardingPath } from './use_card_url_rewrite';
 
 describe('useIntegratrionCardList', () => {
   describe('toOnboardingPath', () => {
