--- x-pack/solutions/observability/plugins/observability_onboarding/public/application/packages_list/lazy.tsx
@@ -1,20 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import type { AvailablePackagesHookType } from '@kbn/fleet-plugin/public';
-import { lazy } from 'react';
-
-export const PackageList = lazy(async () => ({
-  default: await import('@kbn/fleet-plugin/public')
-    .then((module) => module.PackageList())
-    .then((pkg) => pkg.PackageListGrid),
-}));
-
-export const fetchAvailablePackagesHook = (): Promise<AvailablePackagesHookType> =>
-  import('@kbn/fleet-plugin/public')
-    .then((module) => module.AvailablePackagesHook())
-    .then((hook) => hook.useAvailablePackages);
